rename recipePermissionsHelpers -> modDefinitionPermissionsHelpers
