+++ src/activation/useActivateRecipe.ts
@@ -26,7 +26,7 @@ import { getErrorMessage } from "@/errors/errorHelpers";
 import { uninstallRecipe } from "@/store/uninstallUtils";
 import { selectExtensions } from "@/store/extensionsSelectors";
 import { ensurePermissionsFromUserGesture } from "@/permissions/permissionsUtils";
-import { checkRecipePermissions } from "@/modDefinitions/recipePermissionsHelpers";
+import { checkRecipePermissions } from "@/modDefinitions/modDefinitionPermissionsHelpers";
 import { isEmpty } from "lodash";
 import { useCreateDatabaseMutation } from "@/services/api";
 import { isDatabaseField } from "@/components/fields/schemaFields/fieldTypeCheckers";
